A few test cleanups
- Move some tests out of `dask.tests.test_base`. These were put in the
  wrong file.

- Remove old skip decorator in favor of `pytest.mark.skip`.
